requirements.txt
# PDF文档解读智能体 - 依赖包

# Web框架
Flask==2.3.3
Flask-CORS==4.0.0

# PDF处理
pdf2image==1.16.3
Pillow==10.0.1

# HTTP请求
requests==2.31.0

# 中文分词
jieba==0.42.1

# 环境变量管理
python-dotenv==1.0.0

# 工具库
werkzeug==2.3.7

# 开发工具
pytest==7.4.2
pytest-flask==1.2.0